crm/requirements.txt
asgiref==3.3.1
companies==1.0.0
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
